requirements.txt
opencv-python
mediapipe
numpy
pillow
catboost
joblib
tk
